[level 2] Title: 피보나치 수, Time: 0.37 ms, Memory: 73.4 MB -BaekjoonHub

+++ 프로그래머스/2/12945. 피보나치 수/피보나치 수.java
@@ -1,20 +1,20 @@
 import java.util.*;
 
 class Solution {
-    HashMap<Integer, Integer> fibo = new HashMap<>();
+    HashMap<Integer, Long> fibo = new HashMap<>();
     int MOD = 1234567;
 
     public int solution(int n) {
-        fibo.put(0, 0); fibo.put(1, 1);
-        return Fibo(n);
+        fibo.put(0, 0L); fibo.put(1, 1L);
+        return (int) Fibo(n);
     }
 
-    private int Fibo(int n) {
+    private long Fibo(int n) {
         if (fibo.containsKey(n)) return fibo.get(n);
         if (n % 2 == 0) {
-            fibo.put(n, (int)(( (long)Fibo(n / 2) * ((2L * Fibo(n / 2 - 1) + Fibo(n / 2)) % MOD) ) % MOD));
+            fibo.put(n, ((Fibo(n / 2) * ((2 * Fibo(n / 2 - 1) + Fibo(n / 2)) % MOD) ) % MOD));
         } else {
-            fibo.put(n, (int)(( (long)Fibo(n / 2) * Fibo(n / 2) % MOD + (long)Fibo(n / 2 + 1) * Fibo(n / 2 + 1) % MOD ) % MOD));
+            fibo.put(n, ((Fibo(n / 2) * Fibo(n / 2) % MOD + (long)Fibo(n / 2 + 1) * Fibo(n / 2 + 1) % MOD ) % MOD));
         }
         return fibo.get(n);
     }
